Refactor HiveMind into modules
